moved components that represent whole pages into another directory
