+++ examples/wallet/app/pages/ChooseRestoreOrImport.scss
