+++ examples/wallet/app/pages/Home.js
