+++ examples/wallet/app/pages/Index.js
