+++ examples/wallet/app/pages/InputKeys.js
